+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+2016.03.23
+----------
+- Make refenrece between artists and tracks
+- Fix missing artist values in the queue
+
 2016.03.21.1
 -------------
 - Fix: Missing ``source`` for ``uri`` to ``spotify_uri``